# University/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/myapp",
...
  "paths": {
    "/api/v1/course/change-details": {
      "put": {
        "tags": [
          "course-controller"
        ],
        "summary": "updateDetails",
        "description": "This API is used to update existing course",
        "operationId": "updateDetailsUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "courseDTO",
            "description": "courseDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/course/de-register/{courseId}": {
      "delete": {
        "tags": [
          "course-controller"
        ],
        "summary": "deleteDetails",
        "description": "This API is used to delete course",
        "operationId": "deleteDetailsUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "courseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/api/v1/department/change-details": {
      "put": {
        "tags": [
          "department-controller"
        ],
        "summary": "updateDetails",
        "description": "This API is used to update existing department",
        "operationId": "updateDetailsUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "departmentDTO",
            "description": "departmentDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DepartmentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/department/de-register/{departmentName}": {
      "delete": {
        "tags": [
          "department-controller"
        ],
        "summary": "deleteDetails",
        "description": "This API is used to delete existing department ",
        "operationId": "deleteDetailsUsingDELETE_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "departmentName",
            "in": "path",
            "description": "departmentName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/department/details/all": {
      "get": {
        "tags": [
          "department-controller"
        ],
        "summary": "getAllDepartment",
        "description": "This API is used fetch all department details",
        "operationId": "getAllDepartmentUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DepartmentResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/department/details/{departmentName}": {
      "get": {
        "tags": [
          "department-controller"
        ],
        "summary": "getDepartmentDetail",
        "description": "This API is used fetch department detail",
        "operationId": "getDepartmentDetailUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "departmentName",
            "in": "path",
            "description": "departmentName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DepartmentResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/department/register": {
      "post": {
        "tags": [
          "department-controller"
        ],
        "summary": "addDepartment",
        "description": "This API is used to create new department",
        "operationId": "addDepartmentUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "departmentDTO",
            "description": "departmentDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DepartmentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/instructor/change-details": {
      "put": {
        "tags": [
          "instructor-controller"
        ],
        "summary": "updateDetails",
        "description": "This API is used update existing instructor detail",
        "operationId": "updateDetailsUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "instructorDTO",
            "description": "instructorDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InstructorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
    "/api/v1/instructor/de-register/{instructorId}": {
      "delete": {
        "tags": [
          "instructor-controller"
        ],
        "summary": "deleteDetails",
        "description": "This API is used to delete existing instructor",
        "operationId": "deleteDetailsUsingDELETE_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "instructorId",
            "in": "path",
            "description": "instructorId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/api/v1/student/change-details": {
      "put": {
        "tags": [
          "student-controller"
        ],
        "summary": "updateDetails",
        "description": "This API is used to update existing student data",
        "operationId": "updateDetailsUsingPUT_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "studentDTO",
            "description": "studentDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StudentDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
...
    "/api/v1/student/de-register/{studentId}": {
      "delete": {
        "tags": [
          "student-controller"
        ],
        "summary": "deleteDetails",
        "description": "This API is used to delete student record",
        "operationId": "deleteDetailsUsingDELETE_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "studentId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT": [
              "global"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "CourseDTO": {
      "type": "object",
      "required": [
        "departmentName",
        "duration",
        "instructorId",
        "name"
      ],
      "properties": {
        "departmentName": {
          "type": "string"
        },
        "duration": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "Required only in case of update"
        },
        "instructorId": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "DepartmentDTO": {
      "type": "object",
      "required": [
        "location",
        "name"
      ],
      "properties": {
        "location": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "DepartmentResponse": {
      "type": "object",
      "properties": {
        "location": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
...
    "InstructorDTO": {
      "type": "object",
      "required": [
        "departmentName",
        "firstName",
        "headedBy",
        "lastName",
        "phone"
      ],
      "properties": {
        "departmentName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "headedBy": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "Required only in case of update"
        },
        "lastName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
    "Message": {
      "type": "object",
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string",
          "description": "Response Message"
        }
      }
    },
...
    "StudentDTO": {
      "type": "object",
      "required": [
        "firstName",
        "lastName",
        "phone"
      ],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "Required only in case of update"
        },
        "lastName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
...
  }
}