# University/src/main/java/com/my/university/response/DepartmentResponse.java
package com.my.university.response;

public class DepartmentResponse {
    private String name;

    private String location;

    public DepartmentResponse(){

    }
    public DepartmentResponse(String name, String location) {
        this.name = name;
        this.location = location;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }

    @Override
    public String toString() {
        return "Department{" +
                "name='" + name + '\'' +
                ", location='" + location + '\'' +
                '}';
    }
}